fix: make payment with failed order after

+++ src/app/checkout/payment/PaymentButton.tsx
@@ -141,7 +141,7 @@ export function PaymentButton ({
       .select('id, payment_status')
       .eq('user_id', userId)
       .then(({ data }: any) => {
-        if (data[0].payment_status === 'pending...') {
+        if (data.length && data[0].payment_status === 'pending...') {
           supabase
             .from('orders')
             .delete()
